Decruft and add coments.
