+++ great_expectations/data_context/templates.py
@@ -175,157 +175,6 @@ data_docs:
             module_name: great_expectations.render.renderer
             class_name: ExpectationSuitePageRenderer
 
-#       site_store: # where the HTML will be written to (filesystem/S3)
-#         type: filesystem
-#         base_directory: uncommitted/documentation/local_site
-#       validations_store: # where to look for validation results (filesystem/S3)
-#         name: local_validation_result_store
-#       profiling_store: # where to look for profiling results (filesystem/S3)
-#         name: local_validation_result_store
-
-#       datasources: '*' # by default, all datasources
-#       sections:
-#         index:
-#           renderer:
-#             module: great_expectations.render.renderer
-#             class: SiteIndexPageRenderer
-#           view:
-#             module: great_expectations.render.view
-#             class: DefaultJinjaIndexPageView
-#         validations: # if not present, validation results are not rendered
-#           run_id_filter:
-#             ne: profiling
-#           renderer:
-#             module: great_expectations.render.renderer
-#             class: ValidationResultsPageRenderer
-#           view:
-#             module: great_expectations.render.view
-#             class: DefaultJinjaPageView
-#         expectations: # if not present, expectation suites are not rendered
-#           renderer:
-#             module: great_expectations.render.renderer
-#             class: ExpectationSuitePageRenderer
-#           view:
-#             module: great_expectations.render.view
-#             class: DefaultJinjaPageView
-#         profiling: # if not present, profiling results are not rendered
-#           run_id_filter:
-#             eq: profiling
-#           renderer:
-#             module: great_expectations.render.renderer
-#             class: ProfilingResultsPageRenderer
-#           view:
-#             module: great_expectations.render.view
-#             class: DefaultJinjaPageView
-
-#     team_site:
-#       # "team_site" is meant to support the "shared source of truth for a team" use case. 
-#       # By default only the expectations section is enabled.
-#       #  Users have to configure the profiling and the validations sections (and the corresponding validations_store and profiling_store attributes based on the team's decisions where these are stored (a local filesystem or S3). 
-#       # Reach out on Slack (https://tinyurl.com/great-expectations-slack>) if you would like to discuss the best way to configure a team site.
-#       type: SiteBuilder
-#       site_store:
-#         type: filesystem
-#         base_directory: uncommitted/documentation/team_site
-# #      validations_store:
-# #      profiling_store:
-
-#       datasources: '*'
-#       sections:
-#         index:
-#           renderer:
-#             module: great_expectations.render.renderer
-#             class: SiteIndexPageRenderer
-#           view:
-#             module: great_expectations.render.view
-#             class: DefaultJinjaIndexPageView
-#         expectations:
-#           renderer:
-#             module: great_expectations.render.renderer
-#             class: ExpectationSuitePageRenderer
-#           view:
-#             module: great_expectations.render.view
-#             class: DefaultJinjaPageView
-
-# data_docs:
-#   sites:
-#     local_site: # site name
-#     # “local_site” renders documentation for all the datasources in the project from GE artifacts in the local repo. 
-#     # The site includes expectation suites and profiling and validation results from uncommitted directory. 
-#     # Local site provides the convenience of visualizing all the entities stored in JSON files as HTML.
-#       type: SiteBuilder
-#       site_store: # where the HTML will be written to (filesystem/S3)
-#         type: filesystem
-#         base_directory: uncommitted/documentation/local_site
-#       validations_store: # where to look for validation results (filesystem/S3)
-#         name: local_validation_result_store
-#       profiling_store: # where to look for profiling results (filesystem/S3)
-#         name: local_validation_result_store
-
-#       datasources: '*' # by default, all datasources
-#       sections:
-#         index:
-#           renderer:
-#             module: great_expectations.render.renderer
-#             class: SiteIndexPageRenderer
-#           view:
-#             module: great_expectations.render.view
-#             class: DefaultJinjaIndexPageView
-#         validations: # if not present, validation results are not rendered
-#           run_id_filter:
-#             ne: profiling
-#           renderer:
-#             module: great_expectations.render.renderer
-#             class: ValidationResultsPageRenderer
-#           view:
-#             module: great_expectations.render.view
-#             class: DefaultJinjaPageView
-#         expectations: # if not present, expectation suites are not rendered
-#           renderer:
-#             module: great_expectations.render.renderer
-#             class: ExpectationSuitePageRenderer
-#           view:
-#             module: great_expectations.render.view
-#             class: DefaultJinjaPageView
-#         profiling: # if not present, profiling results are not rendered
-#           run_id_filter:
-#             eq: profiling
-#           renderer:
-#             module: great_expectations.render.renderer
-#             class: ProfilingResultsPageRenderer
-#           view:
-#             module: great_expectations.render.view
-#             class: DefaultJinjaPageView
-
-#     team_site:
-#       # "team_site" is meant to support the "shared source of truth for a team" use case. 
-#       # By default only the expectations section is enabled.
-#       #  Users have to configure the profiling and the validations sections (and the corresponding validations_store and profiling_store attributes based on the team's decisions where these are stored (a local filesystem or S3). 
-#       # Reach out on Slack (https://tinyurl.com/great-expectations-slack>) if you would like to discuss the best way to configure a team site.
-#       type: SiteBuilder
-#       site_store:
-#         type: filesystem
-#         base_directory: uncommitted/documentation/team_site
-# #      validations_store:
-# #      profiling_store:
-
-#       datasources: '*'
-#       sections:
-#         index:
-#           renderer:
-#             module: great_expectations.render.renderer
-#             class: SiteIndexPageRenderer
-#           view:
-#             module: great_expectations.render.view
-#             class: DefaultJinjaIndexPageView
-#         expectations:
-#           renderer:
-#             module: great_expectations.render.renderer
-#             class: ExpectationSuitePageRenderer
-#           view:
-#             module: great_expectations.render.view
-#             class: DefaultJinjaPageView
-
 """
 
 PROJECT_TEMPLATE = PROJECT_HELP_COMMENT + "datasources: {}\n" + PROJECT_OPTIONAL_CONFIG_COMMENT
