+++ great_expectations/data_context/types/resource_identifiers.py
@@ -4,6 +4,7 @@ logger = logging.getLogger(__name__)
 from six import string_types
 
 from .base_resource_identifiers import (
+    DataContextKey,
     OrderedDataContextKey,
 )
 
@@ -53,3 +54,22 @@ class ValidationResultIdentifier(OrderedDataContextKey):
     # NOTE: This pattern is kinda awkward. It would be nice to ONLY specify _key_order
     _required_keys = set(_key_order)
     _allowed_keys = set(_key_order)
+
+# TODO: Rename to SiteSectionKey, for consistency
+class SiteSectionIdentifier(DataContextKey):
+    _required_keys = set([
+        "site_section_name",
+        "resource_identifier",
+    ])
+    _allowed_keys = _required_keys
+    _key_types = {
+        "site_section_name" : string_types,
+        "resource_identifier" : DataContextKey,
+        # "resource_identifier", ... is NOT strictly typed, since it can contain any type of ResourceIdentifier
+    }
+
+    def __hash__(self):
+        return hash(self.site_section_name+"::"+self.resource_identifier.to_string())
+
+    def __eq__(self, other):
+        return self.__hash__() == other.__hash__()
